code refactoring - simplifying the css
I have separated the css code so that the main elements are on a single, separated css page that applies to every page consistently. this allows for more space on the secondary css pages that are page specific. this also means that the content will not be confusing to edit throughout the website.