blog page created
I made the blog page and added some additional tweaks to some of the other pages such as the portfolio page and the main css styling page for the portfolio. I also changed the document title on all pages and also added a favicon for my portfolio on all the pages too.